refactor: мутации ячеек перенесены из документа в лист
