+++ client/nuxt.config.ts
@@ -11,7 +11,6 @@ export default defineNuxtConfig({
   alias: {
     tslib: 'tslib/tslib.es6.js'
   },
-  buildDir: 'nuxt-dist',
   /*
   ** Nuxt rendering mode
   ** See https://nuxtjs.org/api/configuration-mode
@@ -69,6 +68,7 @@ export default defineNuxtConfig({
     '@nuxtjs/apollo',
     '@nuxtjs/i18n',
     '@nuxtjs/vuetify',
+    // @ts-ignore
     ['@pinia/nuxt', { disableVuex: false }]
   ],
   apollo: {
